Answer May 10th

+++ java/src/main/java/com/studio3104/leetcode/explore/challenge/y2021/May/day10/Solution.java
@@ -24,8 +24,8 @@ class Solution {
 
             primes.add(i);
 
-            for (int j = i * i; j < n; j += i) {
-                isPrime[j] = false;
+            for (long j = (long) i * i; j < n; j += i) {
+                isPrime[(int) j] = false;
             }
         }
